Updates to use S3 bucket instead of EC2 port forwarding
