+++ backend/pe/Makefile
@@ -161,8 +161,8 @@ run-mailer:
 
 # Invoke peAsmSyncController locally (self-contained pe-worker container, like scans).
 #   make -C backend/pe start build
-#   make -C backend/pe run-asmsync ORGS=DHS
+#   make -C backend/pe run-asmsync PHASE=import_s3 ORGS=DHS COUNT=1
 run-asmsync:
-	ORGS="$(ORGS)" \
+	PHASE="$(PHASE)" ORGS="$(ORGS)" \
 		COUNT="$(COUNT)" \
 		"$(PE_DIR)/tools/run_local_asmsync.sh"
