+++ backend/pe/src/pe_asm/remote_step/asm_sync_remote_helpers/enum_ips_from_subs.py
@@ -36,7 +36,6 @@ def link_ip_from_domain(sub, root_uid, org_uid, data_source, conn):
         (DATE, ip_hash, ip, org_uid, sub, data_source, root_uid, None),
     )
     cur.fetchone()
-    # print(row)
     conn.commit()
     cur.close()
     return 1
