+++ dev.env.example
@@ -206,3 +206,5 @@ FLARE_TENANT_ID=
 # Comma-separated Flare API keys. peScanController validates each key and starts
 # one worker container per valid key (capped by COUNT/taskCount).
 FLARE_API_KEYS=
+
+PE_S3_BUCKET=
